Day 25 - Bad Code To The Rescue
