+++ lib/pages/cart_page.dart
@@ -77,17 +77,24 @@ class _CartListState extends State<_CartList> {
   final _cart = CartModel();
   @override
   Widget build(BuildContext context) {
-    return ListView.builder(
-      itemCount: _cart.items.length,
-      itemBuilder: (context, index) => ListTile(
-        leading: const Icon(Icons.done),
-        trailing: IconButton(
-          icon: const Icon(Icons.remove_circle),
-          onPressed: () {},
-        ),
-        title: _cart.items[index].bookName.text.make(),
-        //"item".text.make(),
-      ),
-    );
+    return _cart.items.isEmpty
+        ? "Nothing to show".text.xl3.makeCentered()
+        : ListView.builder(
+            itemCount: _cart.items.length,
+            itemBuilder: (context, index) => ListTile(
+              leading: const Icon(Icons.done),
+              trailing: IconButton(
+                icon: const Icon(Icons.remove_circle),
+                onPressed: () {
+                  _cart.remove(_cart.items[index]);
+                  setState(() {});
+
+                  //_cart.remove(_cart.items[index]);
+                },
+              ),
+              title: _cart.items[index].bookName.text.make(),
+              //"item".text.make(),
+            ),
+          );
   }
 }
