+++ lib/pages/home_detail.dart
@@ -2,6 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:velocity_x/velocity_x.dart';
 
 import '../models/catalog.dart';
+import '../widgets/home_widgets/add_to_cart.dart';
 
 class HomeDetailPage extends StatelessWidget {
   final Item catalog;
@@ -20,16 +21,20 @@ class HomeDetailPage extends StatelessWidget {
           buttonPadding: EdgeInsets.zero,
           children: [
             "\$${catalog.price}".text.bold.xl4.red800.make(),
-            ElevatedButton(
-              onPressed: () {},
-              style: ButtonStyle(
-                  backgroundColor: MaterialStateProperty.all(
-                    context.theme.buttonColor,
-                  ),
-                  shape: MaterialStateProperty.all(
-                    const StadiumBorder(),
-                  )),
-              child: "Add to cart".text.make(),
+            // ElevatedButton(
+            //   onPressed: () {},
+            //   style: ButtonStyle(
+            //       backgroundColor: MaterialStateProperty.all(
+            //         context.theme.buttonColor,
+            //       ),
+            //       shape: MaterialStateProperty.all(
+            //         const StadiumBorder(),
+            //       )),
+            //   child: "Add to cart".text.make(),
+            // ).wh(150, 50)
+            AddToCart(
+              
+              catalog: catalog,
             ).wh(150, 50)
           ],
         ).p16(),
